+++ docs/sidebars.js
@@ -5,6 +5,6 @@ module.exports = {
     Fetching: ['queries', 'mutations', 'fetch-policies', 'pagination', 'error-handling'],
     Caching: ['cache-configuration', 'cache-interaction', 'garbage-collection'],
     Flutter: ['flutter', 'flutter-operation-widget', 'structuring-queries'],
-    Advanced: ['custom-scalars', 'customization'],
+    Advanced: ['custom-scalars', 'customization', 'isolates'],
   },
 };