+++ packages/ferry/test/typed_links/request_controller_typed_link_test.dart
@@ -1,11 +1,13 @@
 import 'dart:async';
 
 import 'package:async/async.dart';
+import 'package:ferry/ferry.dart';
 import 'package:ferry/src/request_controller_typed_link.dart';
 import 'package:ferry_exec/ferry_exec.dart';
 import 'package:ferry_test_graphql/queries/__generated__/reviews.data.gql.dart';
 import 'package:ferry_test_graphql/queries/__generated__/reviews.req.gql.dart';
 import 'package:ferry_test_graphql/queries/__generated__/reviews.var.gql.dart';
+import 'package:gql_exec/gql_exec.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:test/test.dart';
 
@@ -51,6 +53,7 @@ void main() {
               .whereType<OperationResponse<TData, TVars>>()
               .where((res) => request == res.operationRequest)
               .startWith(OperationResponse(
+                dataSource: DataSource.Link,
                 operationRequest: request,
                 data: dataForRequest(request),
               )),
@@ -215,6 +218,150 @@ void main() {
 
         expect(await queue.hasNext, equals(false));
       });
+
+      test('disposing client emits done', () async {
+        final req1 = req.rebuild((b) => b..requestId = '123');
+
+        final reviews1 = dataForRequest(req1).reviews!;
+
+        final req2 = req1.rebuild((b) => b..vars.offset = 3);
+
+        final reviews2 = reviews1.rebuild(
+          (b) => b..addAll(dataForRequest(req2).reviews!),
+        );
+
+        final req3 = req1.rebuild((b) => b..vars.offset = 6);
+
+        final reviews3 = reviews2.rebuild(
+          (b) => b..addAll(dataForRequest(req3).reviews!),
+        );
+
+        final stream = typedLink.request(req1);
+
+        unawaited(expectLater(
+            stream,
+            emitsInOrder([
+              isA<OperationResponse<GReviewsData, GReviewsVars>>()
+                  .having((p) => p.data!.reviews, 'reviews', reviews1),
+              isA<OperationResponse<GReviewsData, GReviewsVars>>()
+                  .having((p) => p.data!.reviews, 'reviews', reviews2),
+              isA<OperationResponse<GReviewsData, GReviewsVars>>()
+                  .having((p) => p.data!.reviews, 'reviews', reviews3),
+              emitsDone
+            ])));
+
+        await typedLink.request(req2).first;
+        await typedLink.request((req3)).first;
+        await responseController.close();
+        requestController.close();
+
+        await typedLink.dispose();
+      });
+
+      test('cancelling subscription and doing new request resets state',
+          () async {
+        final req1 = req.rebuild((b) => b..requestId = '123');
+
+        final reviews1 = dataForRequest(req1).reviews!;
+
+        final req2 = req1.rebuild((b) => b..vars.offset = 3);
+
+        final reviews2 = reviews1.rebuild(
+          (b) => b..addAll(dataForRequest(req2).reviews!),
+        );
+
+        final req3 = req1.rebuild((b) => b..vars.offset = 6);
+
+        final reviews3 = reviews2.rebuild(
+          (b) => b..addAll(dataForRequest(req3).reviews!),
+        );
+
+        final stream = typedLink.request(req1);
+        var lastResponse;
+
+        final sub = stream.listen((response) {
+          lastResponse = response.data!.reviews;
+        });
+
+        await typedLink.request(req2).first;
+        await typedLink.request((req3)).first;
+
+        await sub.cancel();
+
+        await Future.delayed(Duration.zero);
+
+        expect(lastResponse, reviews3);
+
+        final stream2 = typedLink.request(req1);
+
+        expectLater(
+            stream2,
+            emits(
+              isA<OperationResponse<GReviewsData, GReviewsVars>>()
+                  .having((p) => p.data!.reviews, 'reviews', reviews1),
+            ));
+
+        await typedLink.dispose();
+      });
     });
   });
+
+  group('tests with real client', () {
+    test('real client test', () async {
+      final client = Client(link: _AutoResponderTerminalLink());
+
+      final req1 = req.rebuild((b) => b..requestId = '123');
+
+      final stream = client.request(req1);
+
+      final reviews1 = dataForRequest(req1).reviews!;
+
+      final req2 = req1.rebuild((b) => b..vars.offset = 3);
+
+      final reviews2 = reviews1.rebuild(
+        (b) => b..addAll(dataForRequest(req2).reviews!),
+      );
+
+      expect(
+          stream,
+          emitsInOrder([
+            isA<OperationResponse<GReviewsData, GReviewsVars>>()
+                .having((p) => p.data!.reviews, 'reviews', reviews1),
+            isA<OperationResponse<GReviewsData, GReviewsVars>>()
+                .having((p) => p.data!.reviews, 'reviews', reviews2),
+            emitsDone
+          ]));
+
+      await Future.delayed(Duration.zero);
+
+      client.requestController.add(req2);
+
+      await Future.delayed(Duration.zero);
+
+      await client.dispose();
+    });
+  });
+}
+
+class _AutoResponderTerminalLink extends Link {
+  @override
+  Stream<Response> request(
+    Request req, [
+    forward,
+  ]) =>
+      Stream.value(Response(
+          response: {},
+          data: GReviewsData((b) => b
+            ..reviews.addAll([
+              for (int i = req.variables['offset'] as int;
+                  i <
+                      ((req.variables['offset'] as int) + req.variables['first']
+                          as int);
+                  i++)
+                GReviewsData_reviews(
+                  (b) => b
+                    ..id = '$i'
+                    ..stars = 5,
+                )
+            ])).toJson()));
 }