+++ packages/ferry_flutter/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.6.2-dev.1
+
+ - Update a dependency to the latest release.
+
+## 0.6.2-dev.0
+
+ - **FEAT**: add IsolateClient (#405).
+
+## 0.6.1+2
+
+ - Update a dependency to the latest release.
+
 ## 0.6.1+1
 
  - Update a dependency to the latest release.